Move `Configuration.get*Name` to `DebugLogger`
Target: RAM: 2459 Flash: 55034
Before: RAM: 2947 Flash: 59326
After:  RAM: 2901 Flash: 59040
