+++ Marlin/tests/configuration/configuration_bed_leveling.cpp
@@ -24,22 +24,6 @@ void test_configuration_multiple_bed_leveling_types_invalid() {
   TEST_ASSERT_FALSE(configuration.validate());
 }
 
-void test_configuration_get_bed_leveling_name() {
-  Configuration configuration;
-  TEST_ASSERT_EQUAL_STRING("", configuration.getBedLevelingName());
-}
-
-void test_configuration_get_bed_leveling_name_3_point() {
-  Configuration configuration;
-  configuration.bedLevelingAuto3Point = true;
-  TEST_ASSERT_EQUAL_STRING("3POINT", configuration.getBedLevelingName());
-}
-
-void test_configuration_get_bed_leveling_name_3_point_explicit() {
-  Configuration configuration;
-  TEST_ASSERT_EQUAL_STRING("3POINT", configuration.getBedLevelingName(CONF_BED_LEVELING_TYPE_3POINT));
-}
-
 void test_configuration_abl_but_not_ubl() {
   Configuration configuration;
   TEST_ASSERT_FALSE(configuration.hasAblButNotUbl());
@@ -96,9 +80,6 @@ void test_configuration_bed_leveling() {
   RUN_TEST(test_configuration_get_bed_leveling_type_3_point);
   RUN_TEST(test_configuration_bed_leveling_type_valid);
   RUN_TEST(test_configuration_multiple_bed_leveling_types_invalid);
-  RUN_TEST(test_configuration_get_bed_leveling_name);
-  RUN_TEST(test_configuration_get_bed_leveling_name_3_point);
-  RUN_TEST(test_configuration_get_bed_leveling_name_3_point_explicit);
   RUN_TEST(test_configuration_abl_but_not_ubl);
   RUN_TEST(test_configuration_abl_but_not_ubl_3_point);
   RUN_TEST(test_configuration_abl_but_not_ubl_ubl);
