+++ Marlin/tests/configuration/configuration_machine_type.cpp
@@ -20,22 +20,8 @@ void test_configuration_multiple_machine_types_invalid() {
   TEST_ASSERT_FALSE(configuration.validate());
 }
 
-void test_configuration_get_machine_name_core_xy() {
-  Configuration configuration;
-  configuration.isCartesian = false;
-  configuration.isCore = true;
-  TEST_ASSERT_EQUAL_STRING("Core", configuration.getMachineName());
-}
-
-void test_configuration_get_machine_name_core_xy_explicit() {
-  Configuration configuration;
-  TEST_ASSERT_EQUAL("Core", configuration.getMachineName(CONF_MACHINE_TYPE_CORE));
-}
-
 void test_configuration_machine_type() {
   RUN_TEST(test_configuration_get_machine_type);
   RUN_TEST(test_configuration_get_machine_type_core_xy);
   RUN_TEST(test_configuration_multiple_machine_types_invalid);
-  RUN_TEST(test_configuration_get_machine_name_core_xy);
-  RUN_TEST(test_configuration_get_machine_name_core_xy_explicit);
 }
